fix nature / test4

+++ Client/CartoonWar/CartoonWar/Codes/LowPoly.cpp
@@ -55,7 +55,7 @@ _int CLowPoly::Update_GameObject(const _float& fTimeDelta)
 
 	if (m_eEnviType == ENVITYPE::ENVI_TREE)
 	{
-		if ((int)(m_pTransformCom->Get_Scale().x) != (int)(server->Get_Nature_Scale(m_iLayerIdx)))
+		if (m_pTransformCom->Get_Scale().x != server->Get_Nature_Scale(m_iLayerIdx))
 		{
 			_uint iSize = server->Get_Nature_Scale(m_iLayerIdx);
 			m_pTransformCom->Scaling(iSize, iSize, iSize);
@@ -63,7 +63,7 @@ _int CLowPoly::Update_GameObject(const _float& fTimeDelta)
 	}
 	else if (m_eEnviType == ENVITYPE::ENVI_ROCK)
 	{
-		if ((int)(m_pTransformCom->Get_Scale().x) != (int)(server->Get_Rock_Scale(m_iLayerIdx)))
+		if (m_pTransformCom->Get_Scale().x != server->Get_Rock_Scale(m_iLayerIdx))
 		{
 			_uint iSize = server->Get_Nature_Scale(m_iLayerIdx);
 			m_pTransformCom->Scaling(iSize, iSize, iSize);
